Version 1.2.8
15.07.2001 - Removed unnecessary fstat() system call in sf_ftp and sf_scp1.
16.07.2001 - Function handle_options() and trans_exec() put the quotes for
             file names with spaces or ";" at the wrong place. Fixed.
17.07.2001 - sf_ftp failed to reset the ascii buffer. This could sometimes
             lead to files being send in ASCII mode and not in binary.
18.07.2001 - Implemented "do not remove" option for local directories.
20.07.2001 - The directory option "delete unknown files" WRONGLY deleted
             files in the queue as well. For this reason a new option has
             been added "delete queued files", which will delete old
             queued files. The old option works now as documented.
22.07.2001 - Reduce the timeout to a maximum of 20 second for function
             ftp_get_reply().
25.07.2001 - Did not have enough memory in append functions to store
             the binary zero. Fixed.
             lead to a SIGSEGV. Hope this has been fixed.
             are not separated with an empty line. Thanks to Bellucci
             Massimo for reporting this one.
26.07.2001 - Added function ftp_account() required by some FTP proxies.
29.07.2001 - New dialog show_queue to view and delete files from the
             AFD queue.
30.07.2001 - Function init_sf_burst2() failed to remove empty file
             directories. Fixed.
31.07.2001 - In function check_burst_2() forgot to set protocol.
04.08.2001 - Added to new fields to HOST_CONFIG file to store the status
             and some protocol specific options for this host.
05.08.2001 - New function ftp_idle() to set idle time of remote FTP-server.
             code.
07.08.2001 - Fixed some display bugs in job details window of the
             afd_ctrl dialog.

+++ amg/check_files.c
@@ -254,19 +254,28 @@ check_files(struct directory_entry *p_de,
                /* Generate name for the new file */
                (void)strcpy(ptr, p_dir->d_name);
 
-               if (p_de->flag & IN_SAME_FILESYSTEM)
+               if ((fra[p_de->fra_pos].remove == YES) ||
+                   (fra[p_de->fra_pos].protocol != LOC))
                {
-                  ret = move_file(fullname, tmp_file_dir);
+                  if (p_de->flag & IN_SAME_FILESYSTEM)
+                  {
+                     ret = move_file(fullname, tmp_file_dir);
+                  }
+                  else
+                  {
+                     ret = copy_file(fullname, tmp_file_dir, &stat_buf);
+                     if (unlink(fullname) == -1)
+                     {
+                        (void)rec(sys_log_fd, ERROR_SIGN,
+                                  "Failed to unlink() file %s : %s (%s %d)\n",
+                                  fullname, strerror(errno), __FILE__, __LINE__);
+                     }
+                  }
                }
                else
                {
-                  ret = copy_file(fullname, tmp_file_dir);
-                  if (unlink(fullname) == -1)
-                  {
-                     (void)rec(sys_log_fd, ERROR_SIGN,
-                               "Failed to unlink() file %s : %s (%s %d)\n",
-                               fullname, strerror(errno), __FILE__, __LINE__);
-                  }
+                  /* Leave original files in place. */
+                  ret = copy_file(fullname, tmp_file_dir, &stat_buf);
                }
                if (ret != SUCCESS)
                {
@@ -429,22 +438,31 @@ check_files(struct directory_entry *p_de,
                            *ptr = '\0';
                         }
 
-                        /* Generate name for the new file */
+                        /* Generate name for the new file. */
                         (void)strcpy(ptr, p_dir->d_name);
-                        if (p_de->flag & IN_SAME_FILESYSTEM)
+                        if ((fra[p_de->fra_pos].remove == YES) ||
+                            (fra[p_de->fra_pos].protocol != LOC))
                         {
-                           ret = move_file(fullname, tmp_file_dir);
+                           if (p_de->flag & IN_SAME_FILESYSTEM)
+                           {
+                              ret = move_file(fullname, tmp_file_dir);
+                           }
+                           else
+                           {
+                              ret = copy_file(fullname, tmp_file_dir, &stat_buf);
+                              if (unlink(fullname) == -1)
+                              {
+                                 (void)rec(sys_log_fd, ERROR_SIGN,
+                                           "Failed to unlink() file %s : %s (%s %d)\n",
+                                           fullname, strerror(errno),
+                                           __FILE__, __LINE__);
+                              }
+                           }
                         }
                         else
                         {
-                           ret = copy_file(fullname, tmp_file_dir);
-                           if (unlink(fullname) == -1)
-                           {
-                              (void)rec(sys_log_fd, ERROR_SIGN,
-                                        "Failed to unlink() file %s : %s (%s %d)\n",
-                                        fullname, strerror(errno),
-                                        __FILE__, __LINE__);
-                           }
+                           /* Leave original files in place. */
+                           ret = copy_file(fullname, tmp_file_dir, &stat_buf);
                         }
                         if (ret != SUCCESS)
                         {
